+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Add European Union as funder (for [STRAITS](https://doi.org/10.3030/101094649)) (#497).
 * New example dataset `river_telemetry` is added to the package (#524).
+* New function `get_receiver_logs()` allows retrieving log files/diagnostic information for acoustic receivers (#254).
 * `read_package()` and `write_package()`, to read and write data packages, are reexported functionality from frictionless, see [`frictionless::read_package()`](https://docs.ropensci.org/frictionless/reference/read_package.html) and [`frictionless::write_package()`](https://docs.ropensci.org/frictionless/reference/write_package.html) (#525).
 
 # etn 3.0.0